+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to MiniJinja are documented here.
 ## 2.1.1
 
 - Added `indent` parameter to `tojson` filter.  #546
+- Added `randrange`, `lipsum`, `random`, `cycler` and `joiner` to
+  `minijinja-contrib`.  #547
 
 ## 2.1.0